sparkstral/pyproject.toml
[project]
name = "sparkstral"
version = "0.1.0"
description = "Sparkstral workflow worker"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.0",
    "mistralai>=2.4.2",
    "mistralai-workflows[mistralai]>=3.3.0",
    "pydantic-settings>=2.14.0",
    "tavily-python>=0.7.23",
]

[dependency-groups]
dev = [
    "mypy>=1.20.2",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.11",
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
strict = true
ignore_missing_imports = true
